Check if posts are cached before dispatching
Before we decide to dispatch a fetch request --and to avoid hitting rate-limits,
check state for a post containing the subreddit that matches the requested subreddit.
If there is a match, then comments have already been loaded and there is no need to dispatch another fetch request.

In this particular scenario --since post data is overwritten with every new
dispatch-- this __quick-n-dirty caching__ is only useful if the user clicks the
same subreddit twice, so I am currently contemplating storing the data of each
subreddit a user visits for the duration of the clients session.

In the future I may decide to use RTK Query to handle this, but for now
relevant fetch dispatches will have pre-conditions check if the data
we need is already in the store.