Extract Up/Down Votes to own component
A reddit `Post` component contains a section where a user interacts with
a post with the option(s) to vote, share, or view comments. This area is
called the "Action Bar". The only references to the term "Action Bar" is the
folder named: `src/components/ActionBar` and in the source-code comments
that define the section.

There is a left action-bar and right action-bar; the left contains up/down
votes UI, the right contains social UI (sharing and viewing comments).

As of this commit, the Votes action-bar section has been extracted from
`Post.jsx` into its own component: `Votes.jsx`, with Social (most likely `Social.jsx`) to
follow.