Remove cache-checking condition
Checking if the post is cached in state before dispatching a fetch
request removed from Sidebar component. It now resides in the async
thunk as a thunk-cancelling condition.

This thunk-cancelling condition will only run when a user consecutively
clicks on the same sidebar link eg. tries to load the same subreddit
that's current in state.

In order for this condition to be useful for reducing end-point
requests, requested subreddit data will have to persist per session
which it currently does not do.