Handle overflowing nav-link text
In life, a long subreddit text eg. "r/someLongSubredditTitle" can break
the layout and cause a frontend developer to succumb to ennui and despair.

We, the lovers of order, patronize this demonic tormentor of movable-type by
setting an `overflow: hidden` on the culpable elements' containing box.

-- Media queries for desktop (LG, XL), and desktop MD have also been added
to handle styling in those specific contexts.